test(fields): pin the missing-key half in its own file so it owns the first render
